cloud_functions/openskynet_pos/requirements.txt
functions-framework==3.*
google-cloud-secret-manager
google-cloud-storage
requests